requirements.txt
# Core
pyyaml>=6.0.1
pydantic>=2.6.0

# Ingestion
pandas>=2.2.0
pyarrow>=15.0.0

dlt>=0.4.7
duckdb>=0.10.0

# Optional validation/reporting
great-expectations>=0.18.12
